Refactor BusinessLogic to adhere to SOLID principles
The `UserAuthenticationService` class was violating the Single Responsibility Principle (SRP) and Interface Segregation Principle (ISP) by handling authentication, password management, and security question logic.

This refactoring breaks down the monolithic service into smaller, more cohesive services:
- `AuthenticationService`: Handles user authentication and 2FA.
- `PasswordService`: Manages password changes and recovery.
- `SecurityQuestionService`: Handles the logic for security questions.
- `SecurityPolicyService`: Manages security policy retrieval and updates.

The dependency injection container has been updated to register these new services. The changes were verified by running the existing unit tests, all of which passed.
